requirements.txt
asgiref==3.8.1
Django==5.2.1
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
